refactor(menu): split data class definitions
