+++ src/aroma/navigation/menu_item.py
@@ -1,18 +1,9 @@
 """
-Defines data structures for handling menu actions and menu items.
+Defines data structures for handling menu items.
 """
-from collections.abc import Callable
 from dataclasses import dataclass
 
-
-@dataclass
-class MenuAction:  # type: ignore
-    """
-    Represents a menu action with associated text and an optional action to be
-    executed.
-    """
-    text: str
-    action: Callable[..., None] | None
+from navigation.menu_action import MenuAction
 
 
 @dataclass
